+++ lib/web_translate_it.rb
@@ -7,10 +7,12 @@ module WebTranslateIt
   def self.fetch_translations
     config = Configuration.new
     if config.autofetch?
-      locale= I18n.locale.to_s
+      locale = I18n.locale.to_s
       puts "Looking for #{locale} translations..."
-      response_code = TranslationFile.fetch(config, locale)
-      puts "Done. Response code: #{response_code}"
+      config.files.each do |file|
+        response_code = file.fetch(locale)
+        puts "Done. Response code: #{response_code}"
+      end
     end
   end
 end