+++ history.md
@@ -1,3 +1,7 @@
+=== (In Git)
+
+* Support for multi-files projects
+
 === Version 1.0.0 / 2009-11-02
 
 * Better support for exceptions